[ACC-247] move passkey validation to it's own hook
