+++ src/hooks/useGlobalAccount.ts
@@ -597,25 +597,6 @@ export const useGlobalAccount = () => {
     return polygonAmoy;
   };
 
-  const validatePasskeyAvailability = async () => {
-    // Availability of `window.PublicKeyCredential` means WebAuthn is usable.
-    // `isUserVerifyingPlatformAuthenticatorAvailable` means the feature detection is usable.
-    // `​​isConditionalMediationAvailable` means the feature detection is usable.
-
-    if (!window.PublicKeyCredential) return false;
-    if (!PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable)
-      return false;
-    if (!PublicKeyCredential.isConditionalMediationAvailable) return false;
-
-    // Check if user verifying platform authenticator is available.
-    const results = await Promise.all([
-      PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable(),
-      PublicKeyCredential.isConditionalMediationAvailable(),
-    ]);
-
-    return results.every((r) => r === true);
-  };
-
   useEffect(() => {
     if (session?.user?.email && !organizationInfo) {
       const { email } = session.user;
@@ -641,7 +622,6 @@ export const useGlobalAccount = () => {
     getKernelClient,
     handleOnChainError,
     getNeededDimoAmountForDcx,
-    validatePasskeyAvailability,
   };
 };
